Explicit conversion from int to double
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk/FermaNext@104 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ src/TrussUnit/TrussPivot.cpp
@@ -53,12 +53,12 @@ void TrussPivot::paint ( ren_dynarow& baseRend, double scaleMultX, double scaleM
     QPoint p1, p2;
     p1 = getFirstNode().getPoint ();
     p2 = getLastNode().getPoint ();
-    p1.setX ( p1.x() * scaleMultX + leftWindowIndent );
-    p1.setY ( flipY ? ( trussAreaHeight - p1.y() ) * scaleMultY + topWindowIndent :
-             p1.y() * scaleMultY + topWindowIndent );
-    p2.setX ( p2.x() * scaleMultX + leftWindowIndent );
-    p2.setY ( flipY ? ( trussAreaHeight - p2.y() ) * scaleMultY + topWindowIndent :
-             p2.y() * scaleMultY + topWindowIndent );
+    p1.setX ( int(p1.x() * scaleMultX) + leftWindowIndent );
+    p1.setY ( flipY ? int(( trussAreaHeight - p1.y() ) * scaleMultY) + topWindowIndent :
+             int(p1.y() * scaleMultY) + topWindowIndent );
+    p2.setX ( int(p2.x() * scaleMultX) + leftWindowIndent );
+    p2.setY ( flipY ? int(( trussAreaHeight - p2.y() ) * scaleMultY) + topWindowIndent :
+             int(p2.y() * scaleMultY) + topWindowIndent );
     if ( isHighlighted () )
         drawLine ( ras, solidRend, sl, p1, p2, pivotsWidth + 5, agg::rgba(200, 135, 15, 0.45) ); 
     drawLine ( ras, solidRend, sl, p1, p2, pivotsWidth, agg::rgba(0, 0, 0) );
